move update intervals to parent widget class.
All widgets now need to define an update method, regardless of whether or
not it has an implementation.

If a widget's configuration in `config.yml` has a value for `refresh`,
it will call the `update` method every `refresh` milliseconds.

This should clean up redundant interval setups in the widgets
themselves and provide an easy way for the user to configure whether or
not they want a given widget to update and if so, how often.